Bugfixing the plug-in code
